feat(navbar): move theme switcher into user popover menu
Move the dark/light theme switcher out of the NavBar button group and
into the user avatar popover, above the "User settings" option, since
it's a personal preference.

The mobile drawer keeps the existing button-style switcher. Shared
toggle logic and the animated icon are extracted into useThemeToggle
and ThemeToggleIcon so both the button and menu-item variants reuse them.
