+++ packages/frontend/src/components/NavBar/ThemeSwitcher/index.ts
@@ -1,1 +1,2 @@
 export { ThemeSwitcher } from './ThemeSwitcher';
+export { ThemeSwitcherMenuItem } from './ThemeSwitcherMenuItem';
